Address attack pipeline review comments
