+++ Assets/Scripts/Combat/Actions/Attacks/AttackResultPipeline.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b6eb5f8982843d2bd0b5c79c0dc9008
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
