+++ Assets/Scripts/Creature/Damage.cs
@@ -41,7 +41,7 @@ namespace Game.Creature
                 string damageTypeCapitalized = char.ToUpper(damageValue.DamageType[0]) + damageValue.DamageType.Substring(1);
                 log += " "+dice.numberOfDice+"d"+dice.sidesPerDie+": "+damageValue.DamageAmount+" " + damageTypeCapitalized + ", ";
                 // Group damage by type (string comparison)
-                AddOrMergeDamage(damageInstances, damageValue);
+                damageInstances = AddOrMergeDamage(damageInstances, damageValue);
 
             }
             log += "\n       ";
@@ -50,7 +50,7 @@ namespace Game.Creature
                 // Group damage by type (string comparison)
                 string damageTypeCapitalized = char.ToUpper(damageFlat.DamageType[0]) + damageFlat.DamageType.Substring(1);
                 log += " +"+damageFlat.DamageAmount + " " + damageTypeCapitalized + ", ";
-                AddOrMergeDamage(damageInstances, damageFlat);
+                damageInstances = AddOrMergeDamage(damageInstances, damageFlat);
             }
             // TODO: append traits list?
 
@@ -59,21 +59,22 @@ namespace Game.Creature
             return damageInstances;
         }
 
-        public static void AddOrMergeDamage(List<DamageValue> damageValues, DamageValue damageValue){
-            if (damageValues == null)
-                return;
-
-            int index = damageValues.FindIndex(damage => damage.DamageType == damageValue.DamageType);
+        /// <summary>
+        /// Returns a new damage list with the supplied value merged by damage type, leaving the input list unchanged.
+        /// </summary>
+        public static List<DamageValue> AddOrMergeDamage(IEnumerable<DamageValue> damageValues, DamageValue damageValue){
+            List<DamageValue> merged = damageValues == null ? new List<DamageValue>() : new List<DamageValue>(damageValues);
+            int index = merged.FindIndex(damage => damage.DamageType == damageValue.DamageType);
             if (index >= 0){
-                DamageValue existing = damageValues[index];
+                DamageValue existing = merged[index];
                 existing.DamageAmount += damageValue.DamageAmount;
-                damageValues[index] = existing;
+                merged[index] = existing;
             }
             else{
-                damageValues.Add(damageValue);
+                merged.Add(damageValue);
             }
+            return merged;
         }
-
         public static int SumDamage(List<DamageValue> damageValues){
             int totalDamage = 0;
             foreach (DamageValue dv in damageValues){
